Organizes and adds batman to the charaters page
