+++ src/components/CharacterInfo/CharacterInfo.js
@@ -1,0 +1,7 @@
+export default function CharacterInfo({character}){
+return (
+  <div className="character-info">
+    <h1>{character.name}</h1>
+  </div>
+)
+}
